Finished DynamicBitset.h
DynamicBitset is a class I made to do the job of std::bitset before I knew what it was, and ultimately it ended up being faster than std::bitset on my cpu and my environment. So, I cleaned it up and added a lower space consuming 32-bit version on top of the 64-bit version. I hope you all find it helpful.