DynamicBitset.h Release
DynamicBitset was a project I started to solve the problem of uint64_t values not having enough bits to work as a mask. At the time, I didn't know what std::bitset was and by the time I found out it was when I was benchmarking my project against it.

Ultimately, on my personal computer DynamicBitset performed consistently better than std::bitset, which can be seen in the DynamicBitset Benchmarks folder in the "Recent Projects" repository pinned to my profile.